requirements.txt
absl-py==1.0.0
asttokens==2.0.5
backcall==0.2.0
blessings==1.7
blobfile==1.2.8
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
dill==0.3.4
easydict==1.9
einops==0.4.1
executing==0.8.3
filelock==3.6.0
fonttools==4.31.1
future==0.18.2
google-auth==2.6.2
google-auth-oauthlib==0.4.6
gpustat==0.6.0
grpcio==1.44.0
guided-diffusion==0.0.0
idna==3.3
imageio==2.16.1
importlib-metadata==4.11.3
ipdb==0.13.9
ipython==8.1.1
jedi==0.18.1
joblib==1.1.0
kiwisolver==1.4.0
Markdown==3.3.6
matplotlib==3.5.1
matplotlib-inline==0.1.3
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mpi4py==3.0.3
networkx==2.7.1
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1634095647912/work
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodomex==3.14.1
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scipy==1.8.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
stack-data==0.2.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
thop==0.0.31.post2005241907
tifffile==2022.3.25
timm==0.3.2
toml==0.10.2
torch==1.7.0
torchaudio==0.7.0a0+ac17b64
torchvision==0.8.0
tqdm==4.63.0
traitlets==5.1.1
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.0.3
xmltodict==0.12.0
zipp==3.7.0
